+++ src/core/timer.h
@@ -11,13 +11,13 @@ typedef void (*TimerFunc_t)(int32_t index);
 typedef struct Timer_t {
   int32_t resetValue;
   int32_t ticksLeft;
-  uint8_t index;
+  int32_t index;
 
   TimerFunc_t  callback;
 } Timer_t;
 
 
-struct Timer_t* Timer_Init(int32_t resetValue, TimerFunc_t callback, uint8_t index);
+struct Timer_t* Timer_Init(int32_t resetValue, TimerFunc_t callback, int32_t index);
 void Timer_Cleanup(struct Timer_t* this);
 
 void Timer_Tick(struct Timer_t* this, int32_t cycles);
